Remove obsolete JS Processor
We have moved all the processing to Symbolicator/Rust quite a while ago, and there is currently no code path that uses the old JS Processor.

Time for it to go. RIP.
